Added algorithm for calculating longest path
This algorithm is not the most ideal but it's based on a few things:

1. It doesn't assume to know anything about the weights or structure
of the tree

2. It only looks for connections and doesn't care if they loop. This
is an actual requirement for calculating the longest path in this
game.

As I learn more about trees, graph traversal and function-based LUTs
through Elixir macros and pattern matching, then I expect the
performance of this to improve.

[#34]